Major architecture refactor
