+++ src/api/routes/articles.js
@@ -1,7 +1,7 @@
 import express from 'express';
 import bodyParser from 'body-parser';
 
-import { auth, isAccessible } from '../config/jwt.config';
+import { auth, isAccessible } from '../../config/jwt.config';
 import ArticlesController from '../controllers/articles.controller';
 
 const router = express.Router();
