+++ src/api/routes/images.js
@@ -1,9 +1,7 @@
-
-
 import express from 'express';
 import bodyParser from 'body-parser';
 
-import { auth } from '../config/jwt.config';
+import { auth } from '../../config/jwt.config';
 
 import ImagesController from '../controllers/images.controller';
