+++ src/api/routes/tags.js
@@ -1,9 +1,7 @@
-
-
 import express from 'express';
 import bodyParser from 'body-parser';
 
-import { auth } from '../config/jwt.config';
+import { auth } from '../../config/jwt.config';
 
 import TagsController from '../controllers/tags.controller';
