+++ src/dal/models/blog.js
@@ -1,5 +1,3 @@
-
-
 import mongoose from 'mongoose';
 import { autoIncrement } from 'mongoose-plugin-autoinc';
