+++ src/dal/repositories/user.repository.js
@@ -1,0 +1,19 @@
+const UserRepository = {
+    get: () => {
+
+    },
+    getAll: () => {
+
+    },
+    create: () => {
+
+    },
+    update: () => {
+
+    },
+    delete: (id) => {
+
+    }
+};
+
+export default UserRepository;
